Build without recent coreutils on ARM
ln --relative was introduced in coreutils 8.16 but we can't build that
on ARM because sb2 doesn't see our updated copy of automake.

As getting that working requires time, time that we don't have now we
are going to use the old Mer's coreutils and replace ln --relative with
a custom script. Only for ARM builds.